backend/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
